refactored file structure and added update user api endpoint
