+++ app/controllers/api/users_controller.rb
@@ -19,8 +19,17 @@ class Api::UsersController < ApplicationController
     end
 
     def show 
-        @user = User.find(params[:id].to_i)
+        @user = User.find(params[:id])
         render :show
+    end
+
+    def update 
+        @user = User.find(params[:id])
+        if @user.update(user_params)
+            render :show
+        else 
+            render json: @user.errors.full_messages, status: 401
+        end
     end
 
     private
