+++ frontend/components/user_dashboard/user_dashboard_main/user_dashboard_main.jsx
@@ -3,6 +3,7 @@ import UserDashboardSidebar from "./user_dashboard_sidebar";
 import UserDashboardCenter from "./user_dashboard_center";
 import {Route, Switch} from "react-router-dom";
 import UserShowContainer from "../user_show/user_show_container"
+import {ProtectedRoute} from "../../../util/route_util"
 
 const UserDashboardMain = () => {
     return (
@@ -10,8 +11,8 @@ const UserDashboardMain = () => {
             <UserDashboardSidebar/>
             <UserDashboardCenter />
             <Switch>
-                <Route path="/user-dashboard/users/:userId" component={UserShowContainer}></Route>
-            </Switch> 
+                <ProtectedRoute path="/user-dashboard/users/:userId" component={UserShowContainer} />
+            </Switch>
         </div>
     )
 }
