+++ frontend/components/user_dashboard/user_show/user_show_container.js
@@ -1,6 +1,6 @@
 import {connect} from "react-redux";
 import UserShow from "./user_show";
-import { fetchUser } from "../../../actions/search_actions";
+import { fetchUser } from "../../../actions/user_actions";
 
 const mapStateToProps = (state, ownProps) => ({
     user: state.entities.users[ownProps.match.params.userId],
