+++ frontend/util/user_api_util.js
@@ -1,0 +1,14 @@
+export const fetchUser = (userId) => (
+    $.ajax({
+        method: "GET",
+        url: `/api/users/${userId}`
+    })
+)
+
+export const updateUser = (user) => (
+    $.ajax({
+        method: "PATCH",
+        url: `/api/users/${user.id}`,
+        data: {user}
+    })
+)
